Handle C-style comments better, fix style warnings
When extracting comment text, skip over both `comment-start` and
`comment-start-skip` resp. `comment-end` and `comment-end-skip`.  This
matters for C-style languages, where both of these variables pairs are
in use.

Remove unused dependency on `cl`.

Quote symbols `like' `this', as described in Section "Documentation
Tips" of the elisp manual, to make the byte compiler happy.